+++ plugin/src/test/kotlin/org/jlleitschuh/gradle/ktlint/TestsCommon.kt
@@ -3,7 +3,7 @@ package org.jlleitschuh.gradle.ktlint
 import org.intellij.lang.annotations.Language
 import java.io.File
 
-const val LOWEST_SUPPORTED_GRADLE_VERSION = "4.10"
+const val LOWEST_SUPPORTED_GRADLE_VERSION = "5.4.1"
 
 fun File.buildFile() = resolve("build.gradle")